everything builds and runs on Linux

+++ tools/Editor/MaterialEditor/MaterialEditor.cpp
@@ -37,6 +37,8 @@ THE SOFTWARE.
 #include <QtWidgets/QListView>
 #include <QtWidgets/QVBoxLayout>
 
+#include <cmath>
+
 namespace Teardrop {
 namespace Tools {
 
@@ -125,9 +127,9 @@ MaterialEditor::MaterialEditor(ProjectItem* materialItem, QWidget* parent/* =0 *
 		// figure out grid row count by taking square root of number of expressions; we will 
 		// take the floor of that value for the number of rows, and the ceiling of that value
 		// as the number of columns
-        qreal sqr = sqrtf(float(nExpr));
-        int rows = int(floorf(sqr));
-        int cols = int(ceilf(sqr));
+        qreal sqr = std::sqrt(float(nExpr));
+        int rows = int(std::floor(sqr));
+        int cols = int(std::ceil(sqr));
 
 		int k = 0;
 		qreal x = 0.f;
